re-arrange e2e tests to be in directories.
