+++ tests/jest/mirror/namespaces_git_mirror.test.js
@@ -1,16 +1,25 @@
 import request from 'supertest'
-import common from "./common";
+import common from "../common";
 
-const testNamespace = "test-namespace"
+const testNamespace = "test-git-namespace"
 
 beforeAll(async () => {
     // delete a 'test-namespace' if it's already exit.
     await request(common.config.getDirektivHost()).delete(`/api/namespaces/${testNamespace}?recursive=true`)
 });
 
-describe('Test namespaces crud operations', () => {
-    it(`should create a new namespace`, async () => {
-        const res = await request(common.config.getDirektivHost()).put(`/api/namespaces/${testNamespace}`)
+describe('Test namespace git mirroring', () => {
+    it(`should create a new git mirrored namespace`, async () => {
+        const res = await request(common.config.getDirektivHost())
+            .put(`/api/namespaces/${testNamespace}`)
+            .send({
+                url: "https://github.com/direktiv/direktiv-examples.git",
+                ref: "main",
+                cron: "",
+                passphrase: "",
+                publicKey: "",
+                privateKey: ""
+            })
         expect(res.statusCode).toEqual(200)
         expect(res.body).toMatchObject({
             namespace: {
@@ -23,21 +32,22 @@ describe('Test namespaces crud operations', () => {
         })
     })
 
-    it(`should get the new namespace`, async () => {
-        const res = await request(common.config.getDirektivHost()).get(`/api/namespaces/${testNamespace}/tree`)
+    it(`should get the new git namespace`, async () => {
+        await sleep(5000)
+        const res = await request(common.config.getDirektivHost()).get(`/api/namespaces/${testNamespace}/tree/aws`)
         expect(res.statusCode).toEqual(200)
         expect(res.body).toMatchObject({
             namespace: testNamespace,
         })
     })
 
-    it(`should delete the new namespace`, async () => {
+    it(`should delete the new git namespace`, async () => {
         const res = await request(common.config.getDirektivHost()).delete(`/api/namespaces/${testNamespace}?recursive=true`)
         expect(res.statusCode).toEqual(200)
         expect(res.body).toMatchObject({})
     })
 
-    it(`should get 404 after the new namespace deletion`, async () => {
+    it(`should get 404 after the new git namespace deletion`, async () => {
         const res = await request(common.config.getDirektivHost()).get(`/api/namespaces/${testNamespace}/tree`)
         expect(res.statusCode).toEqual(404)
         expect(res.body).toMatchObject({
@@ -46,3 +56,7 @@ describe('Test namespaces crud operations', () => {
         })
     })
 })
+
+function sleep(time) {
+    return new Promise((resolve) => setTimeout(resolve, time));
+}
